fix: Fixes issue where regions cannot be deleted (#612)
Fixes the issue where the last region in an asset is unable to be deleted. When all the regions on an asset are deleted the asset transitions back into the "Visited" state. Therefore, when the asset metadata is saved is is not persisted back to the file system since it is no longer in the "Tagged" state. The fix is to instead of saving the asset metadata file it should be deleted from the storage provider since the asset metadata file doesn't not provide any additional metadata outside of region data.

Resolves AB#17174